// src/schemas/schemas.ts
export interface TypesOfCustomers {
    type: string;
    percentage: number;
}

export interface SalesByRegion {
    norte: number;
    nordeste: number;
    centroOeste: number;
    sudeste: number;
    sul: number;
}

export interface rankTopCustomers {
    valor: number;
    nome: string;
    estado: string;
    tipo_empresa: string;
    tipo: string;
}